Feat: Hp바 추가

+++ Assets/02.Scripts/Skill/SkillConfig.cs
@@ -6,7 +6,7 @@ public class SkillConfig : ScriptableObject
     public Sprite skillIcon;
     public string skillName;
     public SkillType skillType;
-    public int skillValue;
+    public float skillValue;
     [TextArea(3, 5)]
     public string skillDescription;
     public SkillBase skillBase;
